#10342: Refactor LightweightModule
