+++ models/demos/t3000/mixtral8x7b/tt/mixtral_attention.py
@@ -6,7 +6,7 @@ import torch
 import ttnn
 from models.utility_functions import nearest_32
 from ttnn import ShardTensorToMesh, ReplicateTensorToMesh, ConcatMeshToTensor
-from models.demos.t3000.mixtral8x7b.tt.mixtral_common import LightweightModule
+from models.common.lightweightmodule import LightweightModule
 
 
 class TtMixtralAttention(LightweightModule):
